pyproject.toml
[project]
name = "pm-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "pyyaml",
    "slack-sdk>=3.36.0",
]
